Updated App model validation logic
Marathon UI doesn't allow scale and suspend operations, if cmd string is empty.
Cmd string can be empty if an executor and container.image are provided. Today,
executor and container.image cannot be provided through UI, but v2 API supports it.

This commit introduces an additional check, that verifies if executor and
container.image are provided, if cmd string is empty.