Avoid modifying props directly
Modifying properties directly can have side effects. This commit
treats the this.props.model as the initial value for this.state.model so
that the FormGroupComponents can link the form fields to an owned
internal state rather than having to rely on the parent component.

This fixes https://github.com/mesosphere/marathon/issues/1735