Extract deployment badge store listener
The Marathon component is now only responsible for "polling" the
DeploymentsStore rather than resolving the badge indicator. This also
has the positive side-effect of removing the direct manipulation of the
tabs property that was happening inside Marathon.

The NavTabsComponent now bind directly the DeploymentStore and react to
the changes by updating its own internal state.