Integrate reactjs table component
Refactor the app list using the reactjs table component. This is the
very first commit in a series to update the app list using reactjs
components. It removes a lot of the current features (e.g. icons,
groups, health bar, ...) and only displays the given app data and applies
the selected filters.