Add Dockerfile
Example usage:

  $ docker build .
  $ docker run --rm -it -p 1337:1337 -p 4200:4200 -e CORSPROXY_SERVICE="localhost:4200 $MARATHON_URL" -e MARATHON_API_URL="/$MARATHON_URL/" <docker_build_sha>

Had to use node 4 due to a known issue (see Dockerfile)

decided to vendor corsproxy because I wanted to be able to access
http://localhost:1337/#/apps instead of http://localhost:1337/localhost:4200/#/apps

Changes made to corsproxy including changing the proxy hosts to a
whitelist.

usage:

  $ corsproxy <default_proxy_host> [<any_other_host>...]

This makes it so that http://localhost:1337 will proxy to the first host
in the list.