Add a popover component
This component is used to add small overlays of content. It has some
base styles (similar to the tooltip), but you are free to add custom
classes and adjust the styling. Content is provided as child nodes.
Visibility is controlled with the respective property.
Whenever you display the component, it will check its own position an
determine whether it should expand to the top or bottom.