Fix persistent volume app scale down warning
App ids tend to be rather log, as they include the full group path. Thus
they should not be used in dialog titles and any other place that has
limited space. Adjust the persistent volume app scale down warning as
the dialog title should not include the app id, for the given reason.