Fix docker image causing errors on submit.
When switching to the docker section but not adding any information
this leads to a server error. Also if the user adds a docker image
and then removes it again it will also cause a server error. This
fix removes the docker part of the configuration if no image is provided.

This closes mesosphere/marathon\#3790