Use MARATHON_SCHEME_PORT label to define scheme of a given port
The links in the tasks view use the same scheme as the Marathon UI
web application but in some cases users expose a service with a
different scheme on a given port.

In order to correctly redirect the user clicking on the link of a given
port, one can use MARATHON_SCHEME_PORT and MARATHON_SCHEME_PORT<N>
labels to define the scheme of the all the ports and/or the scheme of port
with index N. The available values for those labels are 'http' and 'https'.

If no label is provided, the scheme is the same as Marathon UI scheme to
avoid breaking compatibility.