Fix URIs not displayed in the configuration of applications
After upgrading from Marathon 1.4.3 to 1.5.6, the display of URIs in the
configuration view was set to "Unspecified" even if some URIs were set
for the application. The problem was due to a change in the schema of
/v2/apps. The "uris" field disappeared and has been completely replaced
by the "fetch" field.

This commit add tests covering an app declaring one and several URIs to
detect the issue and fixes it.

Warning: This fix might not be backward compatible with versions below 1.4.3.